radio responds to any web interface
changed from localhost to 0.0.0.0 for the web host. This allows one to reach the radio with a browser on the LAN by using the IP address. Not just limited to using localhost. 

There are repercussions if you don't use localhost. With http://not-local-host-ip-address for instance, which is what we have, you don't get microphone permissions easy or at all. This is an area of active discussion today - what to do about it, and what use cases are affected by this.