index.html - Removed orphaned Log Level dropdown from Debug Settings config.js - Removed log_level from gathering/populating/test-reset code
This cleans up the logging in two ways. We renamed "System Log" to
"Web Interface Log" because the Status section log file is JavaScript
messages. Since it's all about JavaScript, then we call it Web Interface
Log. The filtering was tested and works.

Second, the unused log level dropdown was removed entirely. We don't
write to a log file. The Verbose/Quiet is the only thing in the Debug
Settings section and it sets the level of information that appears in the
console.